refactor(format): extract analysis fun output adapters
